src/rundir.?: make get_pidfile_name public
get_pidfile_name is used publicly now to store the pidfile in the
config struct. Note that I am currently evaluating whether this information
needs to be in the config struct and it may be made private again if the
struct is modified to no longer store the pidfile. The argument is that
it is just as easy to call getpid(), and the pidfile mostly serves to make
sure the daemon isn't already running.