src/logging.c: experiencing 80s flashback
the people who run these computers suck. trying to get anything
productive done is like trying to work on a 1400 baud modem - not
quite a character every 3 seconds, more like a line every 3 seconds. it
that in order to feel like i'm doing anything with my life, i'm resorting
to the chunkhost web console because we can't have anything but office
on these pieces of shit. i hate you guys so much.